Fix for overlaps not being Node instances
This commit fixes a bug in the overlap algorithm. The function
countIdealOverlaps is used to assign to each Node instance the nodes with
which it overlaps. In the overlap algorithm, the overlapCount of each of the
Node instances is reduced as soon as the node with the most overlap is
removed. However, countIdealOverlaps does not assign Node instances to the
overlaps variable of each node, but Interval result objects. This commit fixes
this behaviour, which ensures the overlapCount is properly reduced and thus
the overlap algorithm works as (most likely) intended.